ref: Move `resetSdkMock` to dedicated file
