+++ packages/replay/test/mocks/resetSdkMock.ts
