+++ packages/replay/test/unit/index-errorSampleRate.test.ts
@@ -6,7 +6,7 @@ import { REPLAY_SESSION_KEY, VISIBILITY_CHANGE_TIMEOUT, WINDOW } from '../../src
 import { ReplayContainer } from './../../src/replay';
 import { PerformanceEntryResource } from './../fixtures/performanceEntry/resource';
 import { BASE_TIMESTAMP, RecordMock } from './../index';
-import { resetSdkMock } from './../mocks';
+import { resetSdkMock } from './../mocks/resetSdkMock';
 import { DomHandler, MockTransportSend } from './../types';
 import { useFakeTimers } from './../utils/use-fake-timers';
